reorganize streeams into streams.rs
